Rename jog_arm.cpp and h files to servo.cpp and h files, fix relvant includes
